python/python_send/photoFromRawLogs/requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
psycopg2-binary==2.8.5
python-dateutil==2.8.1
requests==2.24.0
six==1.15.0
urllib3==1.25.10
